requirements.txt
Django==5.1.3
requests==2.32.3
python-dotenv==1.0.1
django-cors-headers==4.6.0
cryptography==43.0.3
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
